✨ (*): update route-reader to read all possible Express routes
Transformations that aren't directly relevant to the domain are in
`utils.js`. The actual logic of finding all Express routes has also been
extracted to `find-express-routes.js`. The actual
runner/requester/writer remains in `route-reader.js`.

Right now it starts a server after requiring in the app (which in
general will be starting a server itself). Not sure if there's a
better approach, as there doesn't seem to be a direct way to close an
Express server instance without access to the return value of
`app.listen()`.